backend/requirements.txt
djangorestframework-simplejwt
djangorestframework
django
pillow